requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.6.20
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
importlib-metadata==1.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.0
protobuf==3.12.2
six==1.15.0
sty==1.0.0b12
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
